pyproject.toml
[project]
name = "op_importer"
requires-python = ">= 3.13"
scripts = { op_import = "op_importer:main" } 
version = "0.1.0"
dependencies = ["python-dotenv>=1.2.1,<2", "pytest"]

[build-system] 
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
authors = ["Will Usher <will.usher@openenergytransition.org>"]
channels = ["conda-forge"]
platforms = ["linux-64"]
preview = ["pixi-build"]

[tool.pixi.package]
name = "op_importer"
version = "0.1.0"

[tool.pixi.pypi-dependencies]
op_importer = { path = ".", editable = true }

[tool.pixi.package.build]
backend = { name = "pixi-build-python", version = "0.4.*"}

[tool.pixi.tasks]
start = "op_import"

[tool.pixi.dependencies]
requests = ">=2.32.5,<3"
pytest = ">=9.0.2,<10"

[tool.pixi.feature.test.tasks]
test = "pytest"

[tool.pixi.environments]
default = { solve-group = "default" }
test = { features = ["test"], solve-group = "default" }

[dependency-groups]
test = ["pytest"]
